+++ src/calc/judgement.rs
@@ -6,22 +6,6 @@ use crate::*;
 use core::cmp::Ordering;
 use core::ops::RangeBounds;
 
-/// Returns `true` if two ranges are mixable.
-///
-/// To be mixable, all positions in two ranges must be compareable.
-pub(crate) fn is_mixable<RX, RY, T>(rx: &RX, ry: &RY) -> bool
-where
-    RX: ?Sized + RangeBounds<T>,
-    RY: ?Sized + RangeBounds<T>,
-    T: ?Sized + PartialOrd,
-{
-    let p1 = bound(rx.start_bound()).pos();
-    let p2 = bound(rx.end_bound()).pos();
-    let p3 = bound(ry.start_bound()).pos();
-    let p4 = bound(ry.end_bound()).pos();
-    util::is_ordered(&[p1, p2, p3, p4])
-}
-
 /// Returns `true` if two ranges are equal.
 pub(crate) fn is_eq<RX, RY, T>(x: &RX, y: &RY) -> bool
 where
@@ -53,10 +37,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let (ex, ey) = (rx.edges(), ry.edges());
     let broken = rx.is_broken() || ry.is_broken();
@@ -73,10 +53,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let (ex, ey) = (rx.edges(), ry.edges());
     let broken = rx.is_broken() || ry.is_broken();
@@ -106,10 +82,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let broken = rx.is_broken() || ry.is_broken();
     let adj = bound_ref(&ry.end_bound()).adjoins(&rx.start_bound());
@@ -123,10 +95,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let broken = rx.is_broken() || ry.is_broken();
     let adj = bound_ref(&rx.end_bound()).adjoins(&ry.start_bound());
@@ -150,10 +118,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let broken = rx.is_broken() || ry.is_broken();
     let adj = bound_ref(&ry.end_bound()).touches(&rx.start_bound());
@@ -167,10 +131,6 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
-        return false;
-    }
-
     let (rx, ry) = (rv::new(rx), rv::new(ry));
     let broken = rx.is_broken() || ry.is_broken();
     let adj = bound_ref(&rx.end_bound()).touches(&ry.start_bound());
@@ -184,24 +144,21 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !is_mixable(rx, ry) {
+    let (rx, ry) = (&rv::new(rx), &rv::new(ry));
+    if rx.is_broken() || ry.is_broken() {
         return None;
     }
 
-    let (x, y) = (&rv::new(rx), &rv::new(ry));
-    let broken = x.is_broken() || y.is_broken();
-    let x_min = util::choose_min(x.edges().0, x.edges().1);
-    let y_min = util::choose_min(y.edges().0, y.edges().1);
-    let x_max = util::choose_max(x.edges().0, x.edges().1);
-    let y_max = util::choose_max(y.edges().0, y.edges().1);
+    let x_min = util::choose_min(rx.edges().0, rx.edges().1);
+    let y_min = util::choose_min(ry.edges().0, ry.edges().1);
+    let x_max = util::choose_max(rx.edges().0, rx.edges().1);
+    let y_max = util::choose_max(ry.edges().0, ry.edges().1);
 
-    if broken {
-        None
-    } else if calc::is_eq(x, y) {
+    if calc::is_eq(rx, ry) {
         Some(Ordering::Equal)
-    } else if x_max < y_min || x_max == y_min && y.is_cursor() {
+    } else if x_max < y_min || x_max == y_min && ry.is_cursor() {
         Some(Ordering::Less)
-    } else if y_max < x_min || y_max == x_min && x.is_cursor() {
+    } else if y_max < x_min || y_max == x_min && rx.is_cursor() {
         Some(Ordering::Greater)
     } else {
         None
@@ -215,11 +172,11 @@ where
     RY: ?Sized + RangeBounds<T>,
     T: ?Sized + PartialOrd,
 {
-    if !calc::is_mixable(rx, ry) {
+    let (rx, ry) = (rv::new(rx), rv::new(ry));
+    if rx.is_broken() || ry.is_broken() {
         return RangeRel::Undefined;
     }
 
-    let (rx, ry) = (rv::new(rx), rv::new(ry));
     let (ex, ey) = (rx.edges(), ry.edges());
     let xs_vs_ys = ex.0.partial_cmp(&ey.0).unwrap();
     let xs_vs_ye = ex.0.partial_cmp(&ey.1).unwrap();
@@ -228,9 +185,7 @@ where
     let xe_meet_ys = bound_ref(&ex.1.bound()).meets(&ey.0.bound(), ps);
     let xs_meet_ye = bound_ref(&ex.0.bound()).meets(&ey.1.bound(), ps);
 
-    if rx.is_broken() || ry.is_broken() {
-        RangeRel::Undefined
-    } else if xs_vs_ys.is_eq() && xe_vs_ye.is_eq() {
+    if xs_vs_ys.is_eq() && xe_vs_ye.is_eq() {
         RangeRel::Equal
     } else if xe_meet_ys || xs_meet_ye {
         RangeRel::Meets(xe_meet_ys)