+++ src/ranges/range_univ.rs
@@ -1598,18 +1598,19 @@ impl<T> RangeUniv<T> {
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::interval::top!()]
-    #[doc = doc_rrb::side::interval::sub::panics::all!()]
     #[doc = doc_rrb::side::interval::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
     ///
     /// let target = ru::new(30..60);
-    /// let r1 = target.interval(&ru::new(50..70));
-    /// let r2 = target.interval(&ru::new(60..80));
-    /// let r3 = target.interval(&ru::new(70..90));
-    /// assert_eq!(r1, None);
+    /// let r1 = target.interval(&ru::new(10..20));
+    /// let r2 = target.interval(&ru::new(20..30));
+    /// let r3 = target.interval(&ru::new(20..40));
+    /// let r4 = target.interval(&ru::new(70..90));
+    /// assert_eq!(r1, Some(ru::new(20..30)));
     /// assert_eq!(r2, None);
-    /// assert_eq!(r3, Some(ru::new(60..70)));
+    /// assert_eq!(r3, None);
+    /// assert_eq!(r4, Some(ru::new(60..70)));
     /// ```
     pub fn interval(&self, other: &Self) -> Option<RangeUniv<T>>
     where
@@ -1622,21 +1623,22 @@ impl<T> RangeUniv<T> {
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::interval_adv::top!()]
-    #[doc = doc_rrb::side::interval_adv::sub::panics::all!()]
     #[doc = doc_rrb::side::interval_adv::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
     /// use rich_range::*;
     ///
     /// let target = ru::new(30..60);
-    /// let r1 = target.interval_adv(&ru::new(50..70), CursorMode::Off);
-    /// let r2 = target.interval_adv(&ru::new(60..80), CursorMode::Off);
-    /// let r3 = target.interval_adv(&ru::new(60..80), CursorMode::On);
-    /// let r4 = target.interval_adv(&ru::new(70..90), CursorMode::Off);
-    /// assert_eq!(r1, None);
+    /// let r1 = target.interval_adv(&ru::new(10..20), CursorMode::Off);
+    /// let r2 = target.interval_adv(&ru::new(20..30), CursorMode::Off);
+    /// let r3 = target.interval_adv(&ru::new(20..30), CursorMode::On);
+    /// let r4 = target.interval_adv(&ru::new(20..40), CursorMode::Off);
+    /// let r5 = target.interval_adv(&ru::new(70..90), CursorMode::Off);
+    /// assert_eq!(r1, Some(ru::new(20..30)));
     /// assert_eq!(r2, None);
-    /// assert_eq!(r3, Some(ru::new(60..60)));    
-    /// assert_eq!(r4, Some(ru::new(60..70)));
+    /// assert_eq!(r3, Some(ru::new(30..30)));
+    /// assert_eq!(r4, None);
+    /// assert_eq!(r5, Some(ru::new(60..70)));
     /// ```
     pub fn interval_adv(&self, other: &Self, mode: CursorMode) -> Option<RangeUniv<T>>
     where
@@ -1650,7 +1652,6 @@ impl<T> RangeUniv<T> {
     #[doc_on_only]
     #[doc = doc_rrb::side::prod::top!()]
     #[doc = doc_rrb::side::prod::sub::notes::all!()]
-    #[doc = doc_rrb::side::prod::sub::panics::all!()]
     #[doc = doc_rrb::side::prod::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1670,7 +1671,6 @@ impl<T> RangeUniv<T> {
     #[doc_on_only]
     #[doc = doc_rrb::side::enwrap::top!()]
     #[doc = doc_rrb::side::enwrap::sub::notes::all!()]
-    #[doc = doc_rrb::side::enwrap::sub::panics::all!()]
     #[doc = doc_rrb::side::enwrap::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1692,7 +1692,6 @@ impl<T> RangeUniv<T> {
     #[doc_on_only]
     #[doc = doc_rrb::side::union::top!()]
     #[doc = doc_rrb::side::union::sub::notes::all!()]
-    #[doc = doc_rrb::side::union::sub::panics::all!()]
     #[doc = doc_rrb::side::union::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1713,7 +1712,6 @@ impl<T> RangeUniv<T> {
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::diff::top!()]
-    #[doc = doc_rrb::side::diff::sub::panics::all!()]
     #[doc = doc_rrb::side::diff::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1737,7 +1735,6 @@ impl<T> RangeUniv<T> {
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::diff_adv::top!()]
-    #[doc = doc_rrb::side::diff_adv::sub::panics::all!()]
     #[doc = doc_rrb::side::diff_adv::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1996,10 +1993,6 @@ macro_rules! impl_bitand {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2017,7 +2010,6 @@ macro_rules! impl_bitand {
             fn bitand(self, rhs: $($rhsRef)?RangeUniv<T>) -> Self::Output {
                 let rx = util::to_ref!($($lhsRef)?, self);
                 let ry = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(rx, ry), "{}", msg::BOUNDS_ORDERED);
                 calc::closed_prod(rx, ry)
             }
         }
@@ -2039,10 +2031,6 @@ macro_rules! impl_bitor {
                 ///
                 /// Returns the merged ranges of two ranges.
                 ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2060,7 +2048,6 @@ macro_rules! impl_bitor {
             fn bitor(self, rhs: $($rhsRef)?RangeUniv<T>) -> Self::Output {
                 let rx = util::to_ref!($($lhsRef)?, self);
                 let ry = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(rx, ry), "{}", msg::BOUNDS_ORDERED);
                 calc::closed_union(rx, ry)
             }
         }
@@ -2089,10 +2076,6 @@ macro_rules! impl_bitxor {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2110,7 +2093,6 @@ macro_rules! impl_bitxor {
             fn bitxor(self, rhs: $($rhsRef)?RangeUniv<T>) -> Self::Output {
                 let rx = util::to_ref!($($lhsRef)?, self);
                 let ry = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(rx, ry), "{}", msg::BOUNDS_ORDERED);
                 calc::closed_enwrap(rx, ry)
             }
         }
@@ -2216,10 +2198,6 @@ macro_rules! impl_bitand_assign {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2241,7 +2219,6 @@ macro_rules! impl_bitand_assign {
             #[inline]
             fn bitand_assign(&mut self, rhs: $($rhsRef)?RangeUniv<T>) {
                 let rhs = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(self, rhs), "{}", msg::BOUNDS_ORDERED);
                 *self = calc::closed_prod(self, rhs);
             }
         }
@@ -2260,10 +2237,6 @@ macro_rules! impl_bitor_assign {
                 /// Performs the `|=` operation.
                 ///
                 /// Assigns merged range of two ranges.
-                ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
                 ///
                 /// # Examples
                 ///
@@ -2286,7 +2259,6 @@ macro_rules! impl_bitor_assign {
             #[inline]
             fn bitor_assign(&mut self, rhs: $($rhsRef)?RangeUniv<T>) {
                 let rhs = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(self, rhs), "{}", msg::BOUNDS_ORDERED);
                 *self = calc::closed_union(self, &rhs);
             }
         }
@@ -2312,10 +2284,6 @@ macro_rules! impl_bitxor_assign {
                 /// - One range is empty, assigns the other range.
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
-                ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
                 ///
                 /// # Examples
                 ///