+++ src/ranges/range_wrapper.rs
@@ -1548,17 +1548,18 @@ where
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::interval::top!()]
-    #[doc = doc_rrb::side::interval::sub::panics::all!()]
     #[doc = doc_rrb::side::interval::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
     ///
     /// let target = rw::new(30..60);
-    /// let r1 = target.interval(&rw::new(50..70));
-    /// let r2 = target.interval(&rw::new(60..80));
-    /// let r3 = target.interval(&rw::new(70..90));
-    /// assert_eq!(r1, None);
+    /// let r1 = target.interval(&rw::new(10..20));
+    /// let r2 = target.interval(&rw::new(20..30));
+    /// let r3 = target.interval(&rw::new(20..40));
+    /// let r4 = target.interval(&rw::new(70..90));
+    /// assert_eq!(r1, Some(ru::new(20..30)));
     /// assert_eq!(r2, None);
+    /// assert_eq!(r3, None);
     /// assert_eq!(r3, Some(ru::new(60..70)));
     /// ```
     pub fn interval(&self, other: &Self) -> Option<RangeUniv<T>>
@@ -1572,21 +1573,22 @@ where
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::interval_adv::top!()]
-    #[doc = doc_rrb::side::interval_adv::sub::panics::all!()]
     #[doc = doc_rrb::side::interval_adv::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
     /// use rich_range::*;
     ///
     /// let target = rw::new(30..60);
-    /// let r1 = target.interval_adv(&rw::new(50..70), CursorMode::Off);
-    /// let r2 = target.interval_adv(&rw::new(60..80), CursorMode::Off);
-    /// let r3 = target.interval_adv(&rw::new(60..80), CursorMode::On);
-    /// let r4 = target.interval_adv(&rw::new(70..90), CursorMode::Off);
-    /// assert_eq!(r1, None);
+    /// let r1 = target.interval_adv(&rw::new(10..20), CursorMode::Off);
+    /// let r2 = target.interval_adv(&rw::new(20..30), CursorMode::Off);
+    /// let r3 = target.interval_adv(&rw::new(20..30), CursorMode::On);
+    /// let r4 = target.interval_adv(&rw::new(20..40), CursorMode::Off);
+    /// let r5 = target.interval_adv(&rw::new(70..90), CursorMode::Off);
+    /// assert_eq!(r1, Some(ru::new(20..30)));
     /// assert_eq!(r2, None);
-    /// assert_eq!(r3, Some(ru::new(60..60)));    
-    /// assert_eq!(r4, Some(ru::new(60..70)));
+    /// assert_eq!(r3, Some(ru::new(30..30)));
+    /// assert_eq!(r4, None);
+    /// assert_eq!(r5, Some(ru::new(60..70)));
     /// ```
     pub fn interval_adv(&self, other: &Self, mode: CursorMode) -> Option<RangeUniv<T>>
     where
@@ -1600,7 +1602,6 @@ where
     #[doc_on_only]
     #[doc = doc_rrb::side::prod::top!()]
     #[doc = doc_rrb::side::prod::sub::notes::all!()]
-    #[doc = doc_rrb::side::prod::sub::panics::all!()]
     #[doc = doc_rrb::side::prod::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1621,7 +1622,6 @@ where
     #[doc_on_only]
     #[doc = doc_rrb::side::enwrap::top!()]
     #[doc = doc_rrb::side::enwrap::sub::notes::all!()]
-    #[doc = doc_rrb::side::enwrap::sub::panics::all!()]
     #[doc = doc_rrb::side::enwrap::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1644,7 +1644,6 @@ where
     #[doc_on_only]
     #[doc = doc_rrb::side::union::top!()]
     #[doc = doc_rrb::side::union::sub::notes::all!()]
-    #[doc = doc_rrb::side::union::sub::panics::all!()]
     #[doc = doc_rrb::side::union::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1666,7 +1665,6 @@ where
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::diff::top!()]
-    #[doc = doc_rrb::side::diff::sub::panics::all!()]
     #[doc = doc_rrb::side::diff::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1690,7 +1688,6 @@ where
     #[must_use]
     #[doc_on_only]
     #[doc = doc_rrb::side::diff_adv::top!()]
-    #[doc = doc_rrb::side::diff_adv::sub::panics::all!()]
     #[doc = doc_rrb::side::diff_adv::sub::examples::head!()]
     /// ```
     /// use rich_range::prelude::*;
@@ -1906,10 +1903,6 @@ macro_rules! impl_bitand {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -1927,7 +1920,6 @@ macro_rules! impl_bitand {
             fn bitand(self, rhs: $($rhsRef)?RangeWrapper<R, T>) -> Self::Output {
                 let rx = util::to_ref!($($lhsRef)?, self);
                 let ry = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(rx, ry), "{}", msg::BOUNDS_ORDERED);
                 calc::closed_prod(rx, ry)
             }
         }
@@ -1950,10 +1942,6 @@ macro_rules! impl_bitor {
                 /// Performs the `|` operation.
                 ///
                 /// Returns the merged ranges of two ranges.
-                ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
                 ///
                 /// # Examples
                 ///
@@ -2001,10 +1989,6 @@ macro_rules! impl_bitxor {
                 /// - One range is empty, returns the other range.
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
-                ///
-                /// # Panics
-                ///
-                /// Panics if range has unordered position like NaN.
                 ///
                 /// # Examples
                 ///
@@ -2135,10 +2119,6 @@ macro_rules! impl_bitand_assign {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2160,7 +2140,6 @@ macro_rules! impl_bitand_assign {
             #[inline]
             fn bitand_assign(&mut self, rhs: $($rhsRef)?RangeWrapper<R, T>) {
                 let rhs = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(self, rhs), "{}", msg::BOUNDS_ORDERED);
                 *self = calc::closed_prod(self, &rhs);
             }
         }
@@ -2181,10 +2160,6 @@ macro_rules! impl_bitor_assign {
                 /// Performs the `|=` operation.
                 ///
                 /// Assigns merged range of two ranges.
-                ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
                 ///
                 /// # Examples
                 ///
@@ -2207,7 +2182,6 @@ macro_rules! impl_bitor_assign {
             #[inline]
             fn bitor_assign(&mut self, rhs: $($rhsRef)?RangeWrapper<R, T>) {
                 let rhs = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(self, rhs), "{}", msg::BOUNDS_ORDERED);
                 *self = calc::closed_union(self, &rhs);
             }
         }
@@ -2236,10 +2210,6 @@ macro_rules! impl_bitxor_assign {
                 ///
                 /// [dbe]: crate::RichRangeBounds#default-broken-empty
                 ///
-                /// # Panics
-                ///
-                /// Panics if the range has unordered position like NaN.
-                ///
                 /// # Examples
                 ///
                 /// ```
@@ -2261,7 +2231,6 @@ macro_rules! impl_bitxor_assign {
             #[inline]
             fn bitxor_assign(&mut self, rhs: $($rhsRef)?RangeWrapper<R, T>) {
                 let rhs = util::to_ref!($($rhsRef)?, rhs);
-                assert!(calc::is_mixable(self, rhs), "{}", msg::BOUNDS_ORDERED);
                 *self = calc::closed_enwrap(self, &rhs);
             }
         }