+++ tests/it/range_view.rs
@@ -558,14 +558,16 @@ fn interval() {
 #[test]
 fn interval_adv() {
     let target = rv::new(&(30..60));
-    let r1 = target.interval_adv(&rv::new(&(50..70)), CursorMode::Off);
-    let r2 = target.interval_adv(&rv::new(&(60..80)), CursorMode::Off);
-    let r3 = target.interval_adv(&rv::new(&(60..80)), CursorMode::On);
-    let r4 = target.interval_adv(&rv::new(&(70..90)), CursorMode::Off);
-    assert_eq!(r1, None);
+    let r1 = target.interval_adv(&rv::new(&(10..20)), CursorMode::Off);
+    let r2 = target.interval_adv(&rv::new(&(20..30)), CursorMode::Off);
+    let r3 = target.interval_adv(&rv::new(&(20..30)), CursorMode::On);
+    let r4 = target.interval_adv(&rv::new(&(20..40)), CursorMode::Off);
+    let r5 = target.interval_adv(&rv::new(&(70..90)), CursorMode::Off);
+    assert_eq!(r1, Some(ru::new(20..30)));
     assert_eq!(r2, None);
-    assert_eq!(r3, Some(ru::new(60..60)));
-    assert_eq!(r4, Some(ru::new(60..70)));
+    assert_eq!(r3, Some(ru::new(30..30)));
+    assert_eq!(r4, None);
+    assert_eq!(r5, Some(ru::new(60..70)));
 }
 
 #[test]