+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Improve performance
+- Rules for unordered values ([#2][issue2])
 
 ## [0.3.1] - 2026-08-21
 
@@ -63,3 +63,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/nossie531/rich_range/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/nossie531/rich_range/releases/tag/v0.1.0
 [issue1]: https://github.com/nossie531/rich_range/issues/1
+[issue2]: https://github.com/nossie531/rich_range/issues/2